manejo de archivos lectura

+++ Cuento.cpp
@@ -1,5 +1,5 @@
 #include "Cuento.h"
-Cuento::Cuento (string titulo, float tiempo_lectura, int anio, Escritor * autor , string titulo_de_libro)
+Cuento::Cuento (string titulo, int tiempo_lectura, int anio, Escritor * autor , string titulo_de_libro)
 : Lectura(titulo, tiempo_lectura, anio, autor)
 {
     this->titulo_de_libro = titulo_de_libro;
