+++ Lectura.cpp
@@ -4,7 +4,7 @@ using namespace std;
 
 #include "Lectura.h"
 
-Lectura::Lectura (string titulo, float tiempo_lectura, int anio, Escritor * autor ){
+Lectura::Lectura (string titulo, int tiempo_lectura, int anio, Escritor * autor ){
     this->titulo = titulo;
     this->tiempo_lectura = tiempo_lectura;
     this->anio = anio;
@@ -23,13 +23,10 @@ bool Lectura::get_leido(){
 string Lectura::get_titulo(){
     return titulo;
 }
-float Lectura::get_tiempo_lectura(){
+int Lectura::get_tiempo_lectura(){
     return tiempo_lectura;
 }
 
 int Lectura::get_anio(){
     return anio;
 }
-
-
-#include "Lectura.h"
