+++ Novela_historica.cpp
@@ -1,7 +1,7 @@
 #include "Novela_historica.h"
 
 
-Novela_historica::Novela_historica(string titulo, float tiempo_lectura, int anio, Escritor * autor, string genero, char* tema)
+Novela_historica::Novela_historica(string titulo, int tiempo_lectura, int anio, Escritor * autor, string genero, char* tema)
 :Novela(titulo, tiempo_lectura, anio, autor, genero)
 {
     this->tema = tema;
